3575 User logins can now be limited to single session at a time
Normally a user can login to the GUI from unlimited number of clients at
the same time. If you want to enforce your users to be able to login only once
(from one client which means device and browser), you can enable this option.
When the user logs out or is inactive for the configured amount of time, the
session is invalidated automatically and the user has to log in again from the
current or another device.